+++ src/main/java/br/com/senior/crm/http/camel/Utils/Enums/Primitive.java
@@ -1,16 +1,21 @@
 package br.com.senior.crm.http.camel.Utils.Enums;
 
-public enum EntitiesAccount
+public enum Primitive
 {
+    // Account
     ACCOUNT("account"),
     ACCOUNT_DEFINITION("accountDefinition"),
     ACCOUNT_PHONE("accountPhone"),
-    ACCOUNT_ADDRESS("accountAddress")
+    ACCOUNT_ADDRESS("accountAddress"),
+    // Opportunity
+
+    // Base Records
+
     ;
 
-    final String path;
+    public String path;
 
-    EntitiesAccount(String path) {
+    Primitive(String path) {
         this.path = path;
     }
 }